fix issues leaderboardapi
