--- packages/client/src/store/game/GameSelectors.ts
@@ -1,5 +1,0 @@
-import { RootState } from '../index'
-
-const getScore = (state: RootState) => state.game.score
-
-export const gameSelectors = { getScore }
